requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
black==24.4.2
certifi==2024.2.2
click==8.1.7
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
typing_extensions==4.11.0
yarl==1.9.4
